recreate polyclinic.UI
