+++ polyclinic.UI/App.xaml.cs
@@ -1,0 +1,12 @@
+﻿namespace polyclinic.UI
+{
+    public partial class App : Application
+    {
+        public App()
+        {
+            InitializeComponent();
+
+            MainPage = new AppShell();
+        }
+    }
+}
