+++ polyclinic.UI/AppShell.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace polyclinic.UI
+{
+    public partial class AppShell : Shell
+    {
+        public AppShell()
+        {
+            InitializeComponent();
+        }
+    }
+}
