+++ polyclinic.UI/MauiProgram.cs
@@ -1,0 +1,25 @@
+﻿using Microsoft.Extensions.Logging;
+
+namespace polyclinic.UI
+{
+    public static class MauiProgram
+    {
+        public static MauiApp CreateMauiApp()
+        {
+            var builder = MauiApp.CreateBuilder();
+            builder
+                .UseMauiApp<App>()
+                .ConfigureFonts(fonts =>
+                {
+                    fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
+                    fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
+                });
+
+#if DEBUG
+		builder.Logging.AddDebug();
+#endif
+
+            return builder.Build();
+        }
+    }
+}
