+++ polyclinic.UI/Platforms/Android/AndroidManifest.xml
