+++ polyclinic.UI/Platforms/Android/Resources/values/colors.xml
