+++ polyclinic.UI/Platforms/MacCatalyst/Info.plist
