+++ polyclinic.UI/Platforms/Windows/App.xaml
@@ -1,8 +1,8 @@
 ﻿<maui:MauiWinUIApplication
-    x:Class="polyclinic.WinUI.App"
+    x:Class="polyclinic.UI.WinUI.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:maui="using:Microsoft.Maui"
-    xmlns:local="using:polyclinic.WinUI">
+    xmlns:local="using:polyclinic.UI.WinUI">
 
 </maui:MauiWinUIApplication>
