+++ polyclinic.UI/Platforms/iOS/AppDelegate.cs
@@ -1,0 +1,10 @@
+﻿using Foundation;
+
+namespace polyclinic.UI
+{
+    [Register("AppDelegate")]
+    public class AppDelegate : MauiUIApplicationDelegate
+    {
+        protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
+    }
+}
