+++ polyclinic.UI/Platforms/iOS/Info.plist
