+++ polyclinic.UI/Resources/Raw/AboutAssets.txt
