+++ polyclinic.UI/Resources/Styles/Colors.xaml
