+++ polyclinic.UI/Resources/Styles/Styles.xaml
