--- polyclinic/App.xaml.cs
@@ -1,11 +1,0 @@
-﻿namespace polyclinic;
-
-public partial class App : Application
-{
-	public App()
-	{
-		InitializeComponent();
-
-		MainPage = new AppShell();
-	}
-}
