--- polyclinic/AppShell.xaml.cs
@@ -1,9 +1,0 @@
-﻿namespace polyclinic;
-
-public partial class AppShell : Shell
-{
-	public AppShell()
-	{
-		InitializeComponent();
-	}
-}
