--- polyclinic/MauiProgram.cs
@@ -1,45 +1,0 @@
-﻿using Microsoft.Extensions.Logging;
-using CommunityToolkit.Maui;
-using polyclinic.Domain.Abstractions;
-using polyclinic.Persistence.Repository;
-using polyclinic.Application.Abstractions;
-using polyclinic.Application.Services;
-
-namespace polyclinic;
-
-public static class MauiProgram
-{
-	public static MauiApp CreateMauiApp()
-	{
-		var builder = MauiApp.CreateBuilder();
-		builder
-			.UseMauiApp<App>()
-            .UseMauiCommunityToolkit()
-            .ConfigureFonts(fonts =>
-			{
-				fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
-				fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
-			});
-
-		SetupServices(builder.Services);
-
-#if DEBUG
-		builder.Logging.AddDebug();
-#endif
-
-		return builder.Build();
-	}
-	private static void SetupServices(IServiceCollection services)
-	{
-		// Services
-		services.AddSingleton<IUnitOfWork, FakeUnitOfWork>();
-		services.AddSingleton<IAppointmentService, AppointmentService>();
-		services.AddSingleton<IClientService, ClientService>();
-		services.AddSingleton<IDoctorService, DoctorService>();
-
-		// Views
-
-		// ViewModels
-
-	}
-}
